ensure bias levels are subtracted as float32, and that the GMOS chip gap fill-in array is the same dtype as the array it's filling in

+++ geminidr/core/primitives_visualize.py
@@ -119,7 +119,8 @@ class Visualize(PrimitivesBASE):
                                     "{} for display".format(ad.filename))
                         log.fullinfo("Bias levels used: {}".format(str(bias_level)))
                         for ext, bias in zip(ad, bias_level):
-                            ext.subtract(bias if bias is not None else 0)
+                            ext.subtract(np.float32(bias) if bias is not None
+                                         else 0)
                     else:
                         log.warning("Bias level not found for {}; approximate "
                                     "bias will not be removed".format(ad.filename))
